day 11 p1 naive, p2 does obviously not work yet
clearly, we need to work with cycles, as the stones are independent.
As we know 0 -> 1 -> 2024 -> [20, 24] -> [2, 0 , 2, 4] this should be
easy if we actually use our brain instead of cpu :D
it also suffices to calculate the update once per unique number and just
keep track of how often we have this number.